Proyecto/fuentes-proyecto/src/sisinf/requirements.txt
absl-py==1.4.0
addict==2.4.0
alabaster==0.7.12
apturl==0.5.2
argopt==0.8.2
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brlapi==0.8.3
cachetools==5.0.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.0
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
crispy-bootstrap5==0.7
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.6
django-cors-headers==3.14.0
django-crispy-forms==2.1
docutils==0.17.1
duplicity==0.8.21
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exrex==0.11.0
fasteners==0.14.1
filelock==3.8.0
Flask==2.3.3
flatbuffers==23.3.3
future==0.18.2
gast==0.4.0
git-fame==2.0.1
google-api-python-client==1.7.11
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.54.0
gyp==0.1
h5py==3.8.0
httplib2==0.20.2
idna==3.3
imagesize==1.3.0
importlib-metadata==4.6.4
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
keras==2.12.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
msgpack==1.0.3
mss==9.0.1
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.3
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.0.1
pipenv==2022.10.12
platformdirs==2.5.2
protobuf==4.22.3
psutil==5.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAutoGUI==0.9.54
pycairo==1.20.1
pycups==2.0.1
PyDrive==1.3.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyscreenshot==3.1
PyScreeze==0.1.29
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-steam-api==1.2.0
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.1
pyusb==1.2.1.post1
pywhatkit==5.4
pyxdg==0.27
PyYAML==5.4.1
ranger-fm==1.9.3
redis==4.5.1
reportlab==3.6.8
requests==2.31.0
requests-futures==1.0.0
requests-oauthlib==1.3.1
roman==3.3
rsa==4.8
s-tui==1.1.3
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==4.3.2
SQLAlchemy==1.4.31
sqlparse==0.4.3
stem==1.8.1
systemd-python==234
tabulate==0.9.0
termcolor==2.3.0
torrequest==0.1.0
tqdm==4.65.0
tweepy==4.14.0
typing-extensions==3.10.0.2
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
urwid==2.1.2
virtualenv==20.16.5
virtualenv-clone==0.5.7
wadllib==1.3.6
websockets==9.1
Werkzeug==2.3.7
wikipedia==1.4.0
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0

###############################################################

[[source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"

[packages]
psycopg2 = "*"
django = "*"
django-crispy-forms = "*"
crispy-bootstrap5 = "*"
beautifulsoup4 = "*"
python-steam-api = "*"

[dev-packages]

[requires]
python_version = "3.10"
